refactor: move type
Move type into file where it makes more thematic sense
