--- src/packument-version-resolving.ts
@@ -1,9 +1,0 @@
-import { VersionNotFoundError } from "./domain/packument";
-import { PackumentNotFoundError } from "./common-errors";
-
-/**
- * A failed attempt at resolving a packument-version.
- */
-export type ResolvePackumentVersionError =
-  | PackumentNotFoundError
-  | VersionNotFoundError;
